use a hook
