+++ packages/frontend/src/features/sqlRunner/components/visualizations/BarChart.tsx
@@ -1,36 +1,34 @@
-import { type BarChartDataTransformer } from '@lightdash/common';
-import { type BarChartConfig } from '@lightdash/common/src/types/visualizations';
 import EChartsReact from 'echarts-for-react';
 import { type EChartsReactProps } from 'echarts-for-react/lib/types';
-import { memo, useMemo, type FC } from 'react';
+import { memo, type FC } from 'react';
+import { useSelector } from 'react-redux';
+import { type useSqlQueryRun } from '../../hooks/useSqlQueryRun';
+import { useBarChartDataTransformer } from '../../transformers/useBarChartDataTransformer';
 
 type BarChartProps = Omit<EChartsReactProps, 'option'> & {
-    transformer: BarChartDataTransformer;
-    config: BarChartConfig;
+    data: NonNullable<ReturnType<typeof useSqlQueryRun>['data']>;
 };
 
-const BarChart: FC<BarChartProps> = memo(
-    ({ transformer, config, className, ...rest }) => {
-        // TODO: should this just be passed in?
-        const spec = useMemo(
-            () => transformer.getEchartsSpec(config),
-            [transformer, config],
-        );
+const BarChart: FC<BarChartProps> = memo(({ data, className, ...rest }) => {
+    // TODO: fix store type
+    const barChartConfig = useSelector(
+        (state: any) => state.sqlRunner.chartConfig,
+    );
+    const { spec } = useBarChartDataTransformer(data, barChartConfig);
 
-        return (
-            <EChartsReact
-                className={className}
-                option={spec}
-                notMerge
-                opts={{
-                    renderer: 'svg',
-                    width: 'auto',
-                    height: 'auto',
-                }}
-                {...rest}
-            />
-        );
-    },
-);
+    return (
+        <EChartsReact
+            className={className}
+            option={spec}
+            notMerge
+            opts={{
+                renderer: 'svg',
+                width: 'auto',
+                height: 'auto',
+            }}
+            {...rest}
+        />
+    );
+});
 
 export default BarChart;
