+++ packages/frontend/src/features/sqlRunner/store/sqlRunnerSlice.ts
@@ -1,14 +1,42 @@
+import { type BarChartConfig } from '@lightdash/common/src/types/visualizations';
 import type { PayloadAction } from '@reduxjs/toolkit';
 import { createSlice } from '@reduxjs/toolkit';
 
 export interface SqlRunnerState {
     projectUuid: string;
     activeTable: string | undefined;
+    chartConfig: BarChartConfig | undefined;
 }
 
 const initialState: SqlRunnerState = {
     projectUuid: '',
     activeTable: undefined,
+    chartConfig: {
+        metadata: {
+            version: 1,
+        },
+        type: 'barChart',
+        style: {
+            legend: {
+                position: 'top',
+                align: 'start',
+            },
+        },
+        axesConfig: {
+            x: {
+                reference: 'status',
+                label: 'moo',
+            },
+            y: [
+                {
+                    reference: 'total_amount',
+                    position: 'left',
+                    label: 'baz',
+                },
+            ],
+        },
+        seriesConfig: [],
+    },
 };
 
 export const sqlRunnerSlice = createSlice({
